whatsonpizza_classifier/requirements.txt
flask
flask-httpauth
gunicorn
pillow
tornado